Core/Vendor: Vendors - Change datatype uint32 to uint64 on HasEnoughMoney - Fix exploit posted on: http://www.arkania.net/arktracker/view.php?id=18 thanks to MysteriousSoul

+++ src/server/game/Entities/Pet/Pet.cpp
@@ -1643,7 +1643,7 @@ bool Pet::resetTalents (bool no_cost)
         return false;
     }
 
-    uint32 cost = 0;
+    uint64 cost = 0;
 
     if (!no_cost)
     {
