+++ src/server/game/Entities/Player/Player.h
@@ -921,10 +921,7 @@ public:
         return m_spellCastItem != 0;
     }
 
-    uint32 GetMoney () const
-    {
-        return m_money;
-    }
+    uint64 GetMoney() const { return m_money; }
     void SetMoney (uint32 money);
 
     bool IsAccepted () const
@@ -1721,19 +1718,13 @@ public:
         m_weaponChangeTimer = time;
     }
 
-    uint32 GetMoney () const
-    {
-        return GetUInt32Value(PLAYER_FIELD_COINAGE);
-    }
+    uint64 GetMoney() const { return GetUInt64Value (PLAYER_FIELD_COINAGE); }
     void ModifyMoney (int32 d);
-    bool HasEnoughMoney (uint32 amount) const
-    {
-        return (GetMoney() >= amount);
-    }
-    bool HasEnoughMoney (int32 amount) const
+    bool HasEnoughMoney(uint64 amount) const { return (GetMoney() >= amount); }
+    bool HasEnoughMoney(int32 amount) const
     {
         if (amount > 0)
-            return (GetMoney() >= uint32(amount));
+        return (GetMoney() >= (uint32) amount);
         return true;
     }
