+++ src/server/game/Handlers/MailHandler.cpp
@@ -113,7 +113,7 @@ void WorldSession::HandleSendMail (WorldPacket & recv_data)
 
     uint64 reqmoney = cost + money;
 
-    if (!pl->HasEnoughMoney(uint32(reqmoney)))
+    if (!pl->HasEnoughMoney(uint64(reqmoney)))
     {
         pl->SendMailResult(0, MAIL_SEND, MAIL_ERR_NOT_ENOUGH_MONEY);
         return;
@@ -421,7 +421,7 @@ void WorldSession::HandleMailTakeItem (WorldPacket & recv_data)
     }
 
     // prevent cheating with skip client money check
-    if (!pl->HasEnoughMoney(uint32(m->COD)))
+    if (!pl->HasEnoughMoney(uint64(m->COD)))
     {
         pl->SendMailResult(mailId, MAIL_ITEM_TAKEN, MAIL_ERR_NOT_ENOUGH_MONEY);
         return;
